32-bit compile fixes (casts from uint64_t to size_t)

+++ renderdoc/driver/d3d12/d3d12_device.cpp
@@ -774,10 +774,10 @@ bool WrappedID3D12Device::Serialise_MapDataWrite(Serialiser *localSerialiser,
 
     if(SUCCEEDED(hr))
     {
-      memcpy(mapPtr + begin, data, end - begin);
+      memcpy(mapPtr + begin, data, size_t(end - begin));
 
-      range.Begin = begin;
-      range.End = end;
+      range.Begin = (size_t)begin;
+      range.End = (size_t)end;
 
       r->Unmap(sub, &range);
     }
